Refactor Reused UW Retro Funcs Into New Clinical Retro Module
Some functions in the `redcap_det_uw_retrospective.py` module were imported and reused
by the `clinical.py` module to expedite the deployment of PHSKC data processing. This
change moves these shared functions into a new module `clinical_retrospectives.py`
which will hold other shared functions needed to process retrospective data regardless
of source.
